requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
bsf
requests
sqlalchemy
newspaper3k
lxml_html_clean
pymysql
python-dotenv